Refactor search screen to handle queried data
Products data updates were previously handled by the search listview, causing problems on search bar unfocus, as the widget was destroyed and rebuilt. Now the list of product documents is handled directly by the search screen, while the listview child only limits to fire some callbacks on events. This improves the structure and the flexibility of the code regarding this functionality.