Remove FutureBuilder from product screen, enable AR button functionality
Now all the product data is directly passed to product screen, has it was already fetched in all other screens that have a route to it. Unity screen now correctly accepts route arguments. Downloader has been slightly modified to use the whole assetBundle path, as it is stored in that way in the database. Small other refactors to product screen layout and complete the AR button function to correctly open Unity.